Feedback from code review
